requirements.txt
### Installed Packages ###

pika
mysql-connector-python
python-decouple
